* Refactor NetDispatcherBase into its own file. * Add thread names to Listener and Clients (within Server). * Change from Sockets->Streams, which should allow for easier unit testing in the future (there are currently still a few casts to NetworkStream).
